Organize js-scripts and create modules
