+++ js/main.js
@@ -1,46 +1,3 @@
+import {createPosts} from './data.js';
 
-const NOTES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-const NAMES = ['Анастасия', 'Кирилл', 'Надежда', 'Ирина', 'Сергей'];
-const POSTS_COUNT = 25;
-
-function getRandomInteger(min, max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-
-const generateComments = (count) => {
-  const commentsArray = [];
-
-  for (let i = 1; i <= count; i++) {
-    const comment = {
-      id: i,
-      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-      message: getRandomArrayElement(NOTES),
-      name: getRandomArrayElement(NAMES),
-    };
-    commentsArray.push(comment);
-  }
-  return commentsArray;
-};
-
-const posts = [];
-
-for (let i = 1; i <= POSTS_COUNT; i++) {
-  const post = {
-    id: i,
-    url: `photos/${i}.jpg`,
-    description: 'Как же прекрасно это время года!',
-    likes: getRandomInteger(15, 200),
-    comments: generateComments(getRandomInteger(1, 5)),
-  };
-  posts.push(post);
-}
+createPosts();
